Title:
Implement responsive navbar and start main layout

Description:

Implemented a responsive navbar using a Bootstrap example, customized to:
Display a menu of links for large screens.
Collapse into a toggle icon for medium and small screens.
Started the main layout:
Added a responsive container that narrows as the screen width increases.
Created the initial section with a page title and an "Add" button.